Mesh translations fix: now matrix is fully based on 3 vectors - rotation, scale, position, so any translations will be handled much better. Rotation fix - now mesh movement is not based on current mesh rotation - mesh is being moved only acros XYZ axes
