+++ Labs/C++ labs/OpenGLDrawer/src/main.cpp
@@ -17,217 +17,11 @@
 #include "general/Window.h"
 #include "general/EventsHandler.h"
 #include "general/ThreadSafeQueue.h"
-#include "loaders/objLoader.h"
 
-/*TODO: replace with vector; move to variables namespace*/
-/* Attributes array */
-int attrs[] =
-{
-        /* Creates to support texture attributes to load into mainShader */
-        /* 3: defines position - x, y, z */
-        /* 2: stands for u, v - texture coordinates*/
-        /* 3: stands for normal vector*/
-        /* 0 is for terminating "for" loop (stupid, but works) */
-        3,2,3,0
-};
+/* Functions for program */
+#include "functions.h"
 
-/* Axis enum class */
-enum AXIS
-{
-    OX, OY, OZ
-};
-
-/* Namespace for global variables into main.cpp file */
-namespace var
-{
-    /* Shaders */
-    std::unique_ptr<Shader> mainShader;
-    std::unique_ptr<Shader> lightShader;
-    std::unique_ptr<Shader> axisShader;
-    std::unique_ptr<Shader> outlineShader;
-
-    /* Textures */
-    std::shared_ptr<Texture> faceTexture;
-    std::shared_ptr<Texture> hairTexture;
-    std::shared_ptr<Texture> defaultTexture;
-
-    /* Meshes vertex arrays */
-    std::vector<float> faceArray;
-    std::vector<float> hairArray;
-    std::vector<float> cubeArray;
-    std::vector<float> tempArray;
-    std::vector<float> oxArr, oyArr, ozArr;
-
-    std::shared_ptr<Mesh> selectedMesh;
-}
-
-/***
- * Draw meshes
-***/
-void drawMeshes(glm::mat4 &proj_view)
-{
-    for(auto &m : ThreadSafeQueue::getMeshes(MESH))
-    {
-        m->draw(GL_TRIANGLES, var::mainShader, proj_view);
-    }
-}
-
-void outlineMesh(glm::mat4 &proj_view)
-{
-    glStencilFunc(GL_NOTEQUAL, 1, 0xFF);
-    glStencilMask(0x00);
-    glDisable(GL_DEPTH_TEST);
-    var::outlineShader->use();
-    glUniform1f(glGetUniformLocation(var::outlineShader->id, "outlining"), 1.004f);
-
-    var::selectedMesh->draw(GL_TRIANGLES, var::outlineShader, proj_view);
-
-    glStencilMask(0xFF);
-    glStencilFunc(GL_ALWAYS, 0, 0xFF);
-    glEnable(GL_DEPTH_TEST);
-}
-
-/***
- * Draw lights
-***/
-void drawLights(glm::mat4 &proj_view)
-{
-    for(auto &m : ThreadSafeQueue::getMeshes(LIGHT))
-    {
-        m->draw(GL_TRIANGLES, var::lightShader, proj_view);
-    }
-}
-
-/***
- * Draw axes
-***/
-void drawAxes(glm::mat4 &proj_view, const glm::vec3 &location, const glm::vec3 &scale)
-{
-    static bool init = false;
-    for(auto &m : ThreadSafeQueue::getMeshes(AXIS))
-    {
-        m->moveTo(location);
-        m->draw(GL_TRIANGLES, var::axisShader, proj_view);
-    }
-    init = true;
-}
-/***
- * Function for loading all the shaders
-***/
-void loadShaders()
-{
-    /* Load mainShader for drawing */
-    std::cout << "Compiling shaders" << std::endl;
-    load_shader("../res/mainv.glsl", "../res/mainf.glsl", var::mainShader);
-    load_shader("../res/lightv.glsl", "../res/lightf.glsl", var::lightShader);
-    load_shader("../res/axisv.glsl", "../res/axisf.glsl", var::axisShader);
-    load_shader("../res/outlinev.glsl", "../res/outlinef.glsl", var::outlineShader);
-    std::cout << "Shaders compiled" << std::endl;
-}
-
-/***
- * Function for loading all the textures
-***/
-void loadTextures()
-{
-    std::cout << "Loading textures" << std::endl;
-    load_texture("../res/Claire/Face/face.png", var::faceTexture);
-    load_texture("../res/Claire/Hair/hair.png", var::hairTexture);
-    load_texture("../res/img.png", var::defaultTexture);
-    std::cout << "Textures loaded" << std::endl;
-}
-
-/***
- * Function for loading hair mesh
-***/
-void loadMeshHair()
-{
-    std::cout << "Loading Hair mesh" << std::endl;
-    objLoader::loadObjModel("../res/Claire/Hair/hairTotal.obj", var::hairArray);
-    std::cout << "Hair mesh loaded" << std::endl;
-}
-
-/***
- * Function for loading face mesh
-***/
-void loadMeshFace()
-{
-    std::cout << "Loading Face mesh" << std::endl;
-    objLoader::loadObjModel("../res/Claire/Face/faceTotal.obj", var::faceArray);
-    std::cout << "Face mesh loaded" << std::endl;
-}
-
-/***
- * Function for loading sphere mesh
-***/
-void loadMeshSphere()
-{
-    std::cout << "Loading Sphere mesh" << std::endl;
-    objLoader::loadObjModel("../res/Basics/sphere.obj", var::cubeArray);
-    std::cout << "Sphere mesh loaded" << std::endl;
-}
-
-void createAxes()
-{
-    objLoader::loadObjModel("../res/Basics/OX.obj", var::oxArr);
-    objLoader::loadObjModel("../res/Basics/OY.obj", var::oyArr);
-    objLoader::loadObjModel("../res/Basics/OZ.obj", var::ozArr);
-}
-
-
-/***
- * Separate function for mesh loading, so we can do this parallel to main thread
-***/
-void loadMeshData(std::string& path, std::vector<float>& verticesArray)
-{
-    objLoader::loadObjModel(path.c_str(), verticesArray);
-}
-
-/***
- * User can enter path to new file he would like to load (and only that so far)
-***/
-void listen()
-{
-    std::string path;
-    while (!Window::isShouldClose())
-    {
-        std::cin >> path;
-        std::cout << path << std::endl;
-        std::cout << "Start loading mesh " << path << std::endl;
-        std::thread loader(loadMeshData, std::ref(path), std::ref(var::tempArray));
-        loader.join();
-        std::cout << "Mesh " << path << "was loaded with vertex size :" << var::tempArray.size() << std::endl;
-    }
-}
-
-void setUp()
-{
-    std::shared_ptr<Mesh> ox = std::make_shared<Mesh>(var::oxArr.data(), var::oxArr.size() / 8, attrs);
-    std::shared_ptr<Mesh> oy = std::make_shared<Mesh>(var::oyArr.data(), var::oyArr.size() / 8, attrs);
-    std::shared_ptr<Mesh> oz = std::make_shared<Mesh>(var::ozArr.data(), var::ozArr.size() / 8, attrs);
-
-    ox->scale(glm::vec3(10, 10, 10));
-    oy->scale(glm::vec3(10, 10, 10));
-    oz->scale(glm::vec3(10, 10, 10));
-
-    ThreadSafeQueue::push(ox, AXIS);
-    ThreadSafeQueue::push(oy, AXIS);
-    ThreadSafeQueue::push(oz, AXIS);
-
-    std::shared_ptr<Mesh> faceMesh = std::make_shared<Mesh>(var::faceArray.data(), var::faceArray.size() / 8, attrs, var::faceTexture);
-    std::shared_ptr<Mesh> hairMesh = std::make_shared<Mesh>(var::hairArray.data(), var::hairArray.size() / 8, attrs, var::hairTexture);
-    std::shared_ptr<Mesh> lightMesh = std::make_shared<Mesh>(var::cubeArray.data(), var::cubeArray.size() / 8, attrs);
-
-    /* Setting meshes default positions */     /*       x           z           y      */
-    lightMesh->moveTo(glm::vec3(-23.1508, 115.151, 146.096));
-    faceMesh->scale(glm::vec3(100, 100, 100));
-    hairMesh->scale(glm::vec3(100, 100, 100));
-
-    /* Appending meshes vector */
-    ThreadSafeQueue::push(faceMesh, MESH);
-    ThreadSafeQueue::push(hairMesh, MESH);
-    ThreadSafeQueue::push(lightMesh, LIGHT);
-}
+/* TODO: replace functions with custom class with static methods like Tick(), Events(), etc. */
 
 int main(int argc, char ** argv)
 {
@@ -300,7 +94,17 @@ int main(int argc, char ** argv)
     /* Tick event */
     while (!Window::isShouldClose())
     {
+        /* Update selected mesh id */
         max = (int) ThreadSafeQueue::getMeshes(ALL).size() - 1;
+        if (id > max)
+        {
+            id = max;
+        }
+        if (id < min)
+        {
+            id = min;
+        }
+
         auto start = std::chrono::high_resolution_clock::now();
 
         /* Global tick events */
@@ -370,30 +174,31 @@ int main(int argc, char ** argv)
         {
             camera->position -= camera->right * delta * speed;
         }
-        /* Move lights */
+
+        /* Move selected mesh */
         if (EventsHandler::pressed(GLFW_KEY_Y))
         {
-            var::selectedMesh->translate(glm::vec3(0, delta * speed, 0));
+            var::selectedMesh->Translate(glm::vec3(0, delta * speed, 0));
         }
         if (EventsHandler::pressed(GLFW_KEY_I))
         {
-            var::selectedMesh->translate(glm::vec3(0, -delta * speed, 0));
+            var::selectedMesh->Translate(glm::vec3(0, -delta * speed, 0));
         }
         if (EventsHandler::pressed(GLFW_KEY_U))
         {
-            var::selectedMesh->translate(glm::vec3(0, 0, - delta * speed));
+            var::selectedMesh->Translate(glm::vec3(0, 0, -delta * speed));
         }
         if (EventsHandler::pressed(GLFW_KEY_H))
         {
-            var::selectedMesh->translate(glm::vec3(-delta * speed, 0, 0));
+            var::selectedMesh->Translate(glm::vec3(-delta * speed, 0, 0));
         }
         if (EventsHandler::pressed(GLFW_KEY_K))
         {
-            var::selectedMesh->translate(glm::vec3(delta * speed, 0, 0));
+            var::selectedMesh->Translate(glm::vec3(delta * speed, 0, 0));
         }
         if (EventsHandler::pressed(GLFW_KEY_J))
         {
-            var::selectedMesh->translate(glm::vec3(0, 0, delta * speed));
+            var::selectedMesh->Translate(glm::vec3(0, 0, delta * speed));
         }
 
         /* Speed up rotation */
@@ -404,6 +209,11 @@ int main(int argc, char ** argv)
         if (EventsHandler::pressed(GLFW_KEY_DOWN))
         {
             rotationSpeed *= .99F;
+        }
+
+        if(EventsHandler::justPressed(GLFW_KEY_DELETE))
+        {
+            ThreadSafeQueue::remove(var::selectedMesh);
         }
 
         /* Switch mesh */
@@ -419,6 +229,7 @@ int main(int argc, char ** argv)
             }
             var::selectedMesh = ThreadSafeQueue::getMeshes(ALL)[id];
         }
+
         if (EventsHandler::justPressed(GLFW_KEY_LEFT))
         {
             if (id == min)
@@ -435,11 +246,11 @@ int main(int argc, char ** argv)
         /* Scale meshes */
         if (EventsHandler::pressed(GLFW_KEY_KP_ADD))
         {
-            var::selectedMesh->scale(glm::vec3(1.01, 1.01, 1.01));
+            var::selectedMesh->Scale(glm::vec3(1.01, 1.01, 1.01));
         }
         if (EventsHandler::pressed(GLFW_KEY_KP_SUBTRACT))
         {
-            var::selectedMesh->scale(glm::vec3(.99, .99, .99));
+            var::selectedMesh->Scale(glm::vec3(.99, .99, .99));
         }
 
         if (EventsHandler::justPressed(GLFW_KEY_SPACE))
@@ -470,29 +281,16 @@ int main(int argc, char ** argv)
         }
         if (rotate)
         {
-            var::selectedMesh->rotate(currentRotationAxis, delta * rotationSpeed);
-            var::selectedMesh->rotate(currentRotationAxis, delta * rotationSpeed);
+            var::selectedMesh->Rotate(currentRotationAxis * delta * rotationSpeed);
         }
 
-        /* Change color */
-        if (EventsHandler::clicked(GLFW_MOUSE_BUTTON_1))
-        {
-            if(lightColor.z > 0.1F)
-            {
-                lightColor.z -= delta;
-            }
-            else
-            {
-                lightColor.z = 1.0F;
-            }
-        }
-
-
+        var::selectedMesh = ThreadSafeQueue::getMeshes(ALL)[id];
         if (!var::tempArray.empty())
         {
             std::shared_ptr<Mesh> temp = std::make_shared<Mesh>(var::tempArray.data(), var::tempArray.size() / 8, attrs, var::defaultTexture);
-            temp->scale(glm::vec3(100, 100, 100));
+            temp->Scale(glm::vec3(100, 100, 100));
             ThreadSafeQueue::push(temp, MESH);
+            var::selectedMesh = temp;
             var::tempArray.clear();
         }
 
@@ -505,16 +303,18 @@ int main(int argc, char ** argv)
         glStencilFunc(GL_ALWAYS, 1, 0xFF);
         glStencilMask(0xFF);
 
+        /* FIXME: this is stupid, we need some kind of storage for lights, so we can pull info about light from it, so we can apply shader variables into draw method call, not here */
         /* Apply lighting data to mainShader */
         var::mainShader->use();
         var::mainShader->uniform3f("proj_pos", camera->position.x, camera->position.y, camera->position.z);
         var::mainShader->uniform3f("light_pos",
-                                   ThreadSafeQueue::getMeshes(LIGHT)[0]->pos.x,
-                                   ThreadSafeQueue::getMeshes(LIGHT)[0]->pos.y,
-                                   ThreadSafeQueue::getMeshes(LIGHT)[0]->pos.z
+                                   ThreadSafeQueue::getMeshes(LIGHT)[0]->position.x,
+                                   ThreadSafeQueue::getMeshes(LIGHT)[0]->position.y,
+                                   ThreadSafeQueue::getMeshes(LIGHT)[0]->position.z
                                    );
         var::mainShader->uniform3f("light_color", lightColor.r, lightColor.g, lightColor.b);
         drawMeshes(proj_view);
+
 
         /* Update light mesh color to match light color */
         var::lightShader->use();
@@ -524,7 +324,7 @@ int main(int argc, char ** argv)
         /* Draw axes at selected mesh location */
         var::axisShader->use();
         var::axisShader->uniform3f("v_color", lightColor.r, lightColor.g, lightColor.b);
-        drawAxes(proj_view, var::selectedMesh->pos, var::selectedMesh->sc);
+        drawAxes(proj_view, var::selectedMesh->position, var::selectedMesh->scale);
 
         /* Outlining selected object */
         outlineMesh(proj_view);
